--- app/handlers/favorites_create_handler.rb
@@ -1,20 +1,0 @@
-# module Favorites
-# end
-class FavoritesCreateHandler < ApplicationServiceHandler
-  def initialize(user:, article_id:)
-    @user = user
-    @article_id = article_id
-  end
-
-  private
-
-  attr_reader :user, :article_id
-
-  def article
-    @article ||= Article.find_by(id: article_id)
-  end
-
-  def handle
-    FavoriteService.new(user: user, article: article).call
-  end
-end
