+++ app/services/application_service.rb
@@ -1,6 +1,13 @@
 class ApplicationService
+  class ServiceError < StandardError; end
   # サービスを利用するクラスが呼び出すメソッド
   def call
-    raise NotImplementedError.new("You must implement #{self.class}##{__method__}")
+    raise ServiceError unless perform
+  end
+
+  private
+
+  def perform
+    raise NotImplementedError, "You must implement #{self.class}##{__method__}"
   end
 end
